+++ domains/service/src/lib.rs
@@ -3,6 +3,7 @@
 mod core_domain;
 mod rpc;
 mod system_domain;
+mod system_domain_tx_pre_validator;
 
 pub use self::core_domain::{new_full_core, CoreDomainParams, NewFullCore};
 pub use self::system_domain::{new_full_system, FullPool, NewFullSystem};
