+++ test/subspace-test-client/src/lib.rs
@@ -46,10 +46,10 @@ use subspace_farmer_components::farming::audit_sector;
 use subspace_farmer_components::plotting::{plot_sector, PieceGetter, PieceGetterRetryPolicy};
 use subspace_farmer_components::{FarmerProtocolInfo, SectorMetadata};
 use subspace_runtime_primitives::opaque::Block;
+use subspace_service::tx_pre_validator::PrimaryChainTxPreValidator;
 use subspace_service::{FullClient, NewFull};
 use subspace_solving::REWARD_SIGNING_CONTEXT;
 use subspace_transaction_pool::bundle_validator::BundleValidator;
-use subspace_transaction_pool::FraudProofVerifierAndBundleValidator;
 use zeroize::Zeroizing;
 
 /// Subspace native executor instance.
@@ -78,12 +78,8 @@ pub type Backend = sc_service::TFullBackend<Block>;
 pub type FraudProofVerifier =
     subspace_service::FraudProofVerifier<subspace_test_runtime::RuntimeApi, TestExecutorDispatch>;
 
-type TxPreValidator = FraudProofVerifierAndBundleValidator<
-    Block,
-    Client,
-    FraudProofVerifier,
-    BundleValidator<Block, Client>,
->;
+type TxPreValidator =
+    PrimaryChainTxPreValidator<Block, Client, FraudProofVerifier, BundleValidator<Block, Client>>;
 
 /// Run a farmer.
 pub fn start_farmer(new_full: &NewFull<Client, TxPreValidator>) {
