+++ test/subspace-test-service/src/mock.rs
@@ -30,21 +30,18 @@ use std::time;
 use subspace_core_primitives::{Blake2b256Hash, Solution};
 use subspace_runtime_primitives::opaque::Block;
 use subspace_runtime_primitives::{AccountId, Hash};
+use subspace_service::tx_pre_validator::PrimaryChainTxPreValidator;
 use subspace_service::FullSelectChain;
 use subspace_solving::create_chunk_signature;
 use subspace_test_client::{Backend, Client, FraudProofVerifier, TestExecutorDispatch};
 use subspace_test_runtime::{RuntimeApi, SLOT_DURATION};
 use subspace_transaction_pool::bundle_validator::BundleValidator;
-use subspace_transaction_pool::{FraudProofVerifierAndBundleValidator, FullPool};
+use subspace_transaction_pool::FullPool;
 
 type StorageChanges = sp_api::StorageChanges<backend::StateBackendFor<Backend, Block>, Block>;
 
-pub(super) type TxPreValidator = FraudProofVerifierAndBundleValidator<
-    Block,
-    Client,
-    FraudProofVerifier,
-    BundleValidator<Block, Client>,
->;
+pub(super) type TxPreValidator =
+    PrimaryChainTxPreValidator<Block, Client, FraudProofVerifier, BundleValidator<Block, Client>>;
 
 /// A mock Subspace primary node instance used for testing.
 pub struct MockPrimaryNode {
@@ -103,7 +100,7 @@ impl MockPrimaryNode {
             task_manager.spawn_handle(),
             subspace_fraud_proof::PrePostStateRootVerifier::new(client.clone()),
         );
-        let tx_pre_validator = FraudProofVerifierAndBundleValidator::new(
+        let tx_pre_validator = PrimaryChainTxPreValidator::new(
             client.clone(),
             Box::new(task_manager.spawn_handle()),
             proof_verifier.clone(),
